% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\docType{data}
\name{coords}
\alias{coords}
\title{Coords}
\format{
An object of class \code{data.frame} with 267 rows and 3 columns.
}
\usage{
data(coords)
}
\description{
\describe{site by x coordinate by y coordinate}
}
\keyword{datasets}
