% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\docType{data}
\name{detection}
\alias{detection}
\title{Detection}
\format{
An object of class \code{data.frame} with 8010 rows and 10 columns.
}
\usage{
data(detection)
}
\description{
\describe{site by season by survey, for six species.
Also contains the standardized covariate duration.}
}
\keyword{datasets}
