% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeBasis.R
\name{MakeBasis}
\alias{MakeBasis}
\title{This function constructs basis functions.  It assumes coordinates form a metric.}
\usage{
MakeBasis(q = q, model.input)
}
\arguments{
\item{q}{The desired number of basis functions.  Must be an integer greater than or equal to 1.}

\item{model.input}{A list of output created by running the create.data.R function}
}
\value{
A list with
\itemize{
\item K spatial basis functions
\item KtQK which is literally the matrix operation transpose(K) times Q times K, and is the variance of the multivariate #' random effect gamma.
}
}
\description{
This function constructs basis functions.  It assumes coordinates form a metric.
}
