% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiocc-package.R
\docType{package}
\name{multiocc-package}
\alias{multiocc}
\alias{multiocc-package}
\title{multiocc: Fits Multivariate Spatio-Temporal Occupancy Model}
\description{
Spatio-temporal multivariate occupancy models can handle multiple species in occupancy models. This method for fitting such models is described in Hepler and Erhardt (2021) "A spatiotemporal model for multivariate occupancy data" \url{https://onlinelibrary.wiley.com/doi/abs/10.1002/env.2657}.
}
\author{
\strong{Maintainer}: Staci Hepler \email{heplersa@wfu.edu}

Authors:
\itemize{
  \item Rob Erhardt
}

}
\keyword{internal}
