% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy.R
\docType{data}
\name{occupancy}
\alias{occupancy}
\title{Occupancy}
\format{
An object of class \code{data.frame} with 2670 rows and 4 columns.
}
\usage{
data(occupancy)
}
\description{
\describe{site by season with covariate forest (standardized percentage forest cover)
and elevation (standardized).}
}
\keyword{datasets}
